Crate a separate package for the web server
