+++ go.mod
@@ -1,3 +1,3 @@
-module opsapi
+module opsapi-server
 
-go 1.16
+go 1.19
