+++ web_server/folder_path.go.bak
@@ -1,0 +1,19 @@
+// Golang code for printing
+// current working directory
+
+package main
+
+import (
+	"fmt"
+	"os"
+)
+
+func main() {
+
+	// using the function
+	mydir, err := os.Getwd()
+	if err != nil {
+		fmt.Println(err)
+	}
+	fmt.Println(mydir)
+}
