+++ web_server/go.mod
@@ -1,0 +1,3 @@
+module web_server
+
+go 1.19
